Extract main trace building function
TODO: we still have to debug why Cairo AIR tests are failing.
