+++ src/air/example/fibonacci_2_columns.rs
@@ -25,19 +25,16 @@ impl From<AirContext> for Fibonacci2ColsAIR {
     }
 }
 
+pub fn build_main_trace(
+    raw_trace: &[Vec<FieldElement<Stark252PrimeField>>],
+) -> TraceTable<Stark252PrimeField> {
+    TraceTable::new_from_cols(raw_trace)
+}
+
 impl AIR for Fibonacci2ColsAIR {
     type Field = Stark252PrimeField;
-    type RawTrace = Vec<Vec<FieldElement<Self::Field>>>;
     type RAPChallenges = ();
     type PublicInput = ();
-
-    fn build_main_trace(
-        &self,
-        raw_trace: &Self::RawTrace,
-        _public_input: &mut Self::PublicInput,
-    ) -> Result<TraceTable<Self::Field>, ProvingError> {
-        Ok(TraceTable::new_from_cols(raw_trace))
-    }
 
     fn build_auxiliary_trace(
         &self,
