+++ src/cairo_vm/execution_trace.rs
@@ -7,11 +7,33 @@ use super::{
     },
 };
 use crate::{
-    air::trace::TraceTable,
+    air::{
+        cairo_air::air::{
+            PublicInputs, FRAME_DST, FRAME_DST_ADDR, FRAME_INST, FRAME_OP0, FRAME_OP0_ADDR,
+            FRAME_OP1, FRAME_OP1_ADDR, FRAME_PC, OFF_DST, OFF_OP0, OFF_OP1,
+        },
+        trace::TraceTable,
+    },
     cairo_vm::{instruction_flags::CairoInstructionFlags, instruction_offsets::InstructionOffsets},
+    prover::ProvingError,
     FE,
 };
-use lambdaworks_math::field::fields::fft_friendly::stark_252_prime_field::Stark252PrimeField;
+use lambdaworks_math::field::{
+    element::FieldElement,
+    fields::fft_friendly::stark_252_prime_field::Stark252PrimeField,
+    traits::{IsFFTField, IsPrimeField},
+};
+
+pub const MEMORY_COLUMNS: [usize; 8] = [
+    FRAME_PC,
+    FRAME_DST_ADDR,
+    FRAME_OP0_ADDR,
+    FRAME_OP1_ADDR,
+    FRAME_INST,
+    FRAME_DST,
+    FRAME_OP0,
+    FRAME_OP1,
+];
 
 // MAIN TRACE LAYOUT
 // -----------------------------------------------------------------------------------------
@@ -26,6 +48,112 @@ use lambdaworks_math::field::fields::fft_friendly::stark_252_prime_field::Stark2
 //  A                B C  D    E    F   G
 // ├xxxxxxxxxxxxxxxx|x|xx|xxxx|xxxx|xxx|xxx┤
 //
+
+pub fn build_main_trace(
+    raw_trace: &(CairoTrace, CairoMemory),
+    public_input: &mut PublicInputs,
+) -> Result<TraceTable<Stark252PrimeField>, ProvingError> {
+    let mut main_trace = build_cairo_execution_trace(&raw_trace.0, &raw_trace.1);
+
+    // First pad is needed by the prover to validate the program bytecode
+    let first_pad = (public_input.program.len() >> 2) + 1;
+    pad_with_last_row(
+        &mut main_trace,
+        (public_input.program.len() >> 2) + 1,
+        &MEMORY_COLUMNS,
+    );
+
+    let (missing_values, rc_min, rc_max) =
+        get_missing_values_offset_columns(&main_trace, &[OFF_DST, OFF_OP0, OFF_OP1]);
+    public_input.range_check_min = Some(rc_min);
+    public_input.range_check_max = Some(rc_max);
+
+    add_missing_values_to_offsets_column(&mut main_trace, missing_values);
+
+    // The second one is a padding to next power of two
+    let padded_trace_length = (raw_trace.0.steps() + first_pad).next_power_of_two();
+
+    if padded_trace_length < main_trace.n_rows() {
+        return Err(ProvingError::WrongParameter(
+            "Trace length is not large enough.".to_string(),
+        ));
+    }
+
+    let padding = padded_trace_length - main_trace.n_rows();
+    pad_with_last_row(&mut main_trace, padding, &MEMORY_COLUMNS);
+
+    Ok(main_trace)
+}
+
+fn pad_with_last_row<F: IsFFTField>(
+    trace: &mut TraceTable<F>,
+    number_rows: usize,
+    exception_columns: &[usize],
+) {
+    let mut last_row = trace.last_row().to_vec();
+    for excemption_column in exception_columns.iter() {
+        last_row[*excemption_column] = FieldElement::zero();
+    }
+    let mut pad: Vec<_> = std::iter::repeat(&last_row)
+        .take(number_rows)
+        .flatten()
+        .cloned()
+        .collect();
+    trace.table.append(&mut pad);
+}
+
+fn get_missing_values_offset_columns<F>(
+    trace: &TraceTable<F>,
+    columns_indices: &[usize],
+) -> (Vec<FieldElement<F>>, u16, u16)
+where
+    F: IsFFTField + IsPrimeField,
+    u16: From<F::RepresentativeType>,
+{
+    let offset_columns = trace.get_cols(columns_indices).table;
+
+    let mut sorted_offset_representatives: Vec<u16> = offset_columns
+        .iter()
+        .map(|x| x.representative().into())
+        .collect();
+    sorted_offset_representatives.sort();
+
+    let mut all_missing_values: Vec<FieldElement<F>> = Vec::new();
+
+    for window in sorted_offset_representatives.windows(2) {
+        if window[1] != window[0] {
+            let mut missing_range: Vec<_> = ((window[0] + 1)..window[1])
+                .map(|x| FieldElement::from(x as u64))
+                .collect();
+            all_missing_values.append(&mut missing_range);
+        }
+    }
+
+    let multiple_of_three_padding =
+        ((all_missing_values.len() + 2) / 3) * 3 - all_missing_values.len();
+    let padding_element = FieldElement::from(*sorted_offset_representatives.last().unwrap() as u64);
+    all_missing_values.append(&mut vec![padding_element; multiple_of_three_padding]);
+
+    (
+        all_missing_values,
+        sorted_offset_representatives[0],
+        sorted_offset_representatives.last().cloned().unwrap(),
+    )
+}
+
+fn add_missing_values_to_offsets_column<F: IsFFTField>(
+    trace: &mut TraceTable<F>,
+    missing_values: Vec<FieldElement<F>>,
+) {
+    let zeros_left = vec![FieldElement::zero(); OFF_DST];
+    let zeros_right = vec![FieldElement::zero(); trace.n_cols - OFF_OP1 - 1];
+
+    for i in (0..missing_values.len()).step_by(3) {
+        trace.table.append(&mut zeros_left.clone());
+        trace.table.append(&mut missing_values[i..(i + 3)].to_vec());
+        trace.table.append(&mut zeros_right.clone());
+    }
+}
 
 /// Receives the raw Cairo trace and memory as outputted from the Cairo VM and returns
 /// the trace table used to feed the Cairo STARK prover.
@@ -351,6 +479,10 @@ fn rows_to_cols<const N: usize>(rows: &[[FE; N]]) -> Vec<Vec<FE>> {
 
 #[cfg(test)]
 mod test {
+    use lambdaworks_math::field::element::FieldElement;
+
+    use crate::air::cairo_air::air::{OFF_DST, OFF_OP1};
+
     use super::*;
 
     #[test]
@@ -838,5 +970,63 @@ mod test {
         ]);
 
         assert_eq!(execution_trace.cols(), expected_trace.cols());
+    }
+    #[test]
+    fn test_fill_range_check_values() {
+        let columns = vec![
+            vec![FieldElement::from(1); 3],
+            vec![FieldElement::from(4); 3],
+            vec![FieldElement::from(7); 3],
+        ];
+        let expected_col = vec![
+            FieldElement::from(2),
+            FieldElement::from(3),
+            FieldElement::from(5),
+            FieldElement::from(6),
+            FieldElement::from(7),
+            FieldElement::from(7),
+        ];
+        let table = TraceTable::<Stark252PrimeField>::new_from_cols(&columns);
+
+        let (col, rc_min, rc_max) = get_missing_values_offset_columns(&table, &[0, 1, 2]);
+        assert_eq!(col, expected_col);
+        assert_eq!(rc_min, 1);
+        assert_eq!(rc_max, 7);
+    }
+
+    #[test]
+    fn test_add_missing_values_to_offsets_column() {
+        let mut main_trace = TraceTable::<Stark252PrimeField> {
+            table: (0..34 * 2).map(FieldElement::from).collect(),
+            n_cols: 34,
+        };
+        let missing_values = vec![
+            FieldElement::from(1),
+            FieldElement::from(2),
+            FieldElement::from(3),
+            FieldElement::from(4),
+            FieldElement::from(5),
+            FieldElement::from(6),
+        ];
+        add_missing_values_to_offsets_column(&mut main_trace, missing_values);
+
+        let mut expected: Vec<_> = (0..34 * 2).map(FieldElement::from).collect();
+        expected.append(&mut vec![FieldElement::zero(); OFF_DST]);
+        expected.append(&mut vec![
+            FieldElement::from(1),
+            FieldElement::from(2),
+            FieldElement::from(3),
+        ]);
+        expected.append(&mut vec![FieldElement::zero(); 34 - OFF_OP1 - 1]);
+        expected.append(&mut vec![FieldElement::zero(); OFF_DST]);
+        expected.append(&mut vec![
+            FieldElement::from(4),
+            FieldElement::from(5),
+            FieldElement::from(6),
+        ]);
+        expected.append(&mut vec![FieldElement::zero(); 34 - OFF_OP1 - 1]);
+        assert_eq!(main_trace.table, expected);
+        assert_eq!(main_trace.n_cols, 34);
+        assert_eq!(main_trace.table.len(), 34 * 4);
     }
 }
